requirements.txt
# Requisitos para el servidor de parqueo ESP32
# Instalar con: pip install -r requirements.txt

# Para manejo de imágenes (opcional)
Pillow>=9.0.0

# Para desarrollo y testing (opcional)
pytest>=7.0.0
